Fixes ich_test.js
There is a minor templating bug in ich_test.js.  In particular,
when passing data to Mustache, one should not nest the value
for a partial under a subdictionary.  This works in Mustache
only because the templates are logic-less; they will look for
all instances of a particular variable and render them appropriately
in the given context.

This is an issue for mustache-based templating systems like
Handlebars for whom nesting is considered.